requirements.txt
fastapi==0.90.1
pydantic==1.10.4
uvicorn==0.20.0
black==23.1.0
sqlalchemy==2.0.3
python-dotenv==0.21.0
psycopg2-binary==2.9.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
httpx==0.23.3
alembic==1.9.3
email-validator==1.3.1
dnspython==2.3.0
jinja2==3.1.2